make entire row clickable + shared component
